[ice] monitor ICE connection for close event
Previously, RTCIceTransport only changed to the "failed" state when a call to
`_recv()` or `_send()` failed. This was fine as long as there was a running
RTCDtlsTransport because it calls `_recv()` in a loop. However if the DTLS
layer has shutdown cleanly, the state change at the ICE layer was never
detected.

This change leverage the new Connection.get_event() coroutine to detect a
failure at the ICE level regardless of whether any data is being sent or
received.

Fixes: #469