[tests] make async tests coroutines
Running the test suite on Python 3.10 triggers a large number of
warnings due to our extensive use of our helper:

```
def run(coro):
    return asyncio.get_event_loop().run_until_complete(coro)
```

The problem is twofold:

- `asyncio.get_event_loop()` will no longer create an event loop in
  future version of Python

- we are alternating between executing code inside and outside of an
  event loop

We solve this by making our test methods coroutines, and adding a
decorator which spins up a single-use loop to run the test. Once we
drop support for Python 3.7 we can remove the decorator and use the
`IsolatedAsyncioTestCase` class which is available in Python 3.8 and up.